Add files via upload
This project was a group effort, this was the final assignment for our CSE 130 Programming with C/C++ class. 
As a group member I provided ideas and feedback, to my classmates. 
I was responsible for the welcome message for our game "Who Wants to be a Millionaire?" 
Below is a short description of my code: 
From line 80 to 90 we have the print statements that will explain how the game “Who Wants to be A Millionaire?” works to the player. The only variable used here is the int variable Num_Questions which contains the number 15, because the game will have a total of 15 questions. At the end of the instructions in line 89 we prompt the user to press Enter to start the program. To get the user input we have the function get() in line 90. After the user presses Enter the game will begin and the user will be prompted to answer the first question.